% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovar.R
\name{auto.var}
\alias{auto.var}
\title{Fit best VAR model to multivariate time series}
\usage{
auto.var(y, max.p = 6, ic = c("SC", "HQ", "AIC", "FPE"), seasonal = TRUE)
}
\arguments{
\item{y}{A multivariate time series}

\item{max.p}{Determines the highest lag order for lag length selection 
according to the choosen ic.}

\item{ic}{Information criterion to be used in model selection.}

\item{seasonal}{If FALSE, restricts search to models without 
seasonal dummies.}
}
\value{
A list with class attribute 'autovar' holding the following elements: 

  \item{\code{fit}}{modelo estimado da class \code{varest}.}
  \item{\code{d}}{ordem de diferenciacao das series.}
  \item{\code{y}}{series das variaveis originais.}
  \item{\code{x}}{series das variaveis diferenciadas \code{d} vezes.}
  \item{\code{ic}}{criterios de informacao do modelo selecionado.}
}
\description{
Returns best ARIMA model according to either SC, HQ, AIC or FPE value.
The function conducts a search over possible model within the order 
constraints provided.
}

