% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descricao}
\alias{descricao}
\title{Descricao das series macroeconomicas usadas}
\format{Um data frame com a descricao das series 
macroeconomicas usadas}
\source{
Elaborada pelos autores
}
\usage{
descricao
}
\description{
Descricao das series macroeconomicas usadas
}
\keyword{datasets}

