% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{di}
\alias{di}
\title{Estima e Prever com Indice de Difussao (di)}
\usage{
di(yh, yt, f, h, m, p)
}
\arguments{
\item{yh}{(ts) serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{(ts) preditor \eqn{y_{t-j+1}}}

\item{f}{(ts) fatores comuns estimados \eqn{F_{t-j+1}}}

\item{h}{horizonte de previsao}

\item{m}{numero de defasagens dos fatores}

\item{p}{numero de defasagens de yt}
}
\value{
uma lista com: fit(dyn): o modelo estimado e 
fcast(ts): valor previsto para \eqn{yh_{t+h}}
}
\description{
Estima uma regressao de yh sobre k fatores e seus m lags e 
sobre yt e seus p lags
}
\details{
yh igual a yt ou a alguma transformacao de yt, esse modelo
de previsao se baseia em Stock e Watson 1998
}

