% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{di.selec}
\alias{di.selec}
\title{Seleciona o modelo di}
\usage{
di.selec(yh, yt, f, h, k, m, p)
}
\arguments{
\item{yh}{serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{preditor \eqn{y_{t-j+1}}}

\item{f}{fatores comuns estimados \eqn{F_{t-j+1}}}

\item{h}{horizonte de previsao}

\item{k}{numero maximo de fatores}

\item{m}{numero maximo de defasagens dos fatores}

\item{p}{numero maximo de defasagens de yt}
}
\value{
best.fit o modelo selecionado dyn
}
\description{
Seleciona o modelo Indice de Difussao com menor BIC
}

