% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{focus}
\alias{focus}
\title{Expectativas de inflacao da Focus}
\format{Um data frame com 168 observacoes de 2002.1 a 2015.12 para a 
expectativas do mercado para o IPCA mensal e para IPCA em 12 meses, 
ambas para um horizonte de 12 meses a frente}
\source{
BCB
}
\usage{
focus
}
\description{
Expectativas de inflacao da Focus
}
\keyword{datasets}

