% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{hard.test}
\alias{hard.test}
\title{Funcao aplica teste t}
\usage{
hard.test(x, dados, h = 12)
}
\arguments{
\item{x}{vetor a ser testado}

\item{dados}{data frame com y e seus quatro lags}

\item{h}{horizonte de previsao}
}
\value{
pvalue com o valor p
}
\description{
calcula o valor p do vetor x na regressao dos dados y=lags(y) + x
em que y esta em dados
}

