% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{hard.threshold}
\alias{hard.threshold}
\title{Selecao de preditores tagerteds}
\usage{
hard.threshold(y, x, h = 12, alpha = 0.05)
}
\arguments{
\item{y}{(ts) serie a ser prevista}

\item{x}{(mts) matriz de preditores sem a variavel target a ser prevista}

\item{h}{(num) numero do horizonte das previsoes fora da amostra}

\item{alpha}{(num) o nivel de significancia critico}
}
\value{
\code{sele} (vector) nome das colunas de x que foram selecionadas
}
\description{
Seleciona as variaveis apresentam poder preditivo 
de acordo com Bai and Ng (2008)
}

