% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macroseries}
\alias{macroseries}
\title{Series macroeconomicas do Brasil}
\format{Um data frame com 217 observacoes de 1996.1 a 2014.2 e 122 variaveis
macroeconomicas para o Brasil}
\source{
Ipeadata e BCB
}
\usage{
macroseries
}
\description{
Series macroeconomicas do Brasil
}
\keyword{datasets}

