% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{outsample.di}
\alias{outsample.di}
\title{Previsao fora da amostra com di}
\usage{
outsample.di(yh, yt, x, k = 1, m = 3, p = 3, n, h = 12)
}
\arguments{
\item{yh}{serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{preditor \eqn{y_{t-j+1}}}

\item{x}{(mts) matriz de preditores}

\item{k}{numero maximo de fatores}

\item{m}{numero maximo de defasagens dos fatores}

\item{p}{numero maximo de defasagens de yt}

\item{n}{numero de previsoes fora da amostra}

\item{h}{horizonte de previsao}
}
\value{
lista contendo \code{fcast} (ts) valores previstos; \code{model} 
(dyn) modelo estimado no fim da amostra
}
\description{
Realiza previsoes fora da amostra para yh com base no 
modelo di selecionado por BIC
}

