% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{outsample.ditp}
\alias{outsample.ditp}
\title{Previsao fora da amostra com di e preditores targeted}
\usage{
outsample.ditp(yh, yt, x, k = 1, m = 3, p = 3, h = 12, n)
}
\arguments{
\item{yh}{serie a ser prevista \eqn{y_{t+h}^h}}

\item{yt}{preditor \eqn{y_{t-j+1}}}

\item{x}{(mts) matriz de preditores}

\item{k}{numero maximo de fatores}

\item{m}{numero maximo de defasagens dos fatores}

\item{p}{numero maximo de defasagens de yt}

\item{h}{horizonte de previsao}

\item{n}{numero de previsoes fora da amostra}
}
\value{
lista contendo \code{fcast} (ts) valores previstos; \code{model} 
(dyn) modelo estimado no fim da amostra; \code{sele} (charc) vetor com o
nome do preditores selecionados
}
\description{
Realiza previsoes fora da amostra para yh com base no 
modelo DI selecionado por BIC e preditores targeted
}

