% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovar.R
\name{outsample.var}
\alias{outsample.var}
\title{Previsao fora da amostra com VAR}
\usage{
outsample.var(y, max.p = 6, ic = c("SC", "HQ", "AIC", "FPE"),
  seasonal = TRUE, h, k)
}
\arguments{
\item{y}{A multivariate time series}

\item{max.p}{Determines the highest lag order for lag length selection 
according to the choosen ic.}

\item{ic}{Information criterion to be used in model selection.}

\item{seasonal}{If FALSE, restricts search to models without 
seasonal dummies.}

\item{h}{horizonte de previsao}

\item{k}{numero de previsoes fora da amostra}
}
\value{
uma lista com
fcast(ts): previsao pontual e intervalos de predicao usando arima 
outdate(num): vetor contendo periodo fora da amostra 
fit(Arima): ultimo modelo arima estimado
}
\description{
outsample.var e usada para gerar previsoes fora da amostra
usando modelo VAR estimado automaticamente com a funcao
auto.var
}

