% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{pc.stand}
\alias{pc.stand}
\title{Estimacao dos fatores com componentes principais}
\usage{
pc.stand(x, norm = TRUE)
}
\arguments{
\item{x}{matriz de preditores}

\item{norm}{logical indicando quando os dados sao padronizados.
Padrao TRUE}
}
\value{
lambda(matrix): loadings; factors(mts): fatores estimado;
autovalor(vector): autovalores de x'x
}
\description{
Estima os fatores usando componentes principais a 
partir da matriz X'X da series padronizadas
}
\details{
Estimacao dos factors com componentes principais segundo 
Stock e Watson (1998). Se norm=TRUE as colunas de X sao 
padronizada para ter media zero e desvio padrao 1
}

