% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/di.R
\name{pc.target}
\alias{pc.target}
\title{Estimacao dos fatores targeteds}
\usage{
pc.target(x, y, h = 12, norm = TRUE)
}
\arguments{
\item{x}{(mts) matriz de preditores}

\item{y}{(ts) variavel a ser prevista}

\item{h}{(int) numero do horizonte das previsoes}

\item{norm}{logical indicando quando os dados sao padronizados. Padrao TRUE}
}
\value{
lambda(matrix) loadings;  factor(ts) fator sintetico 
estimado; autovalor(vector) autovalores de (1/(N*(T-h)))x'x
}
\description{
Estima os fatores usando componentes principais a 
partir da matriz X'X da series padronizadas
}
\details{
Estimacao de um fator sintetico com componentes principais 
segundo Dias et.al. (2010). Se norm=TRUE as colunas de X sao 
padronizadas para ter media zero e desvio padrao 1
}

