% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autovar.R
\name{predict.autovar}
\alias{predict.autovar}
\title{Predict method for objects of class autovar}
\usage{
\method{predict}{autovar}(object, ..., n.ahead = 12, ci = 0.95)
}
\arguments{
\item{object}{An object of class 'autovar'; generated by auto.var().}

\item{...}{Currently not used.}

\item{n.ahead}{An integer specifying the number of forecast steps.}

\item{ci}{The forecast confidence interval.}
}
\value{
A list with class attribute 'varprd' holding the following 
elements: 

  \item{\code{fcst}}{A list of matrices per variable containing 
  the forecasted values with lower and upper bounds 
  as well as the confidence interval.}
  \item{model}{The estimated VAR object for auto.var().}
}
\description{
Forecating a VAR object of class 'autovar' with confidence bands
}

