% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawmacroseries}
\alias{rawmacroseries}
\title{Series macroeconomicas do Brasil}
\format{Um data frame com 217 observacoes de 1996.1 a 2014.2 e 122 variaveis
macroeconomicas para o Brasil com as series brutas, nao transformadas}
\source{
Ipeadata e BCB
}
\usage{
rawmacroseries
}
\description{
Series macroeconomicas do Brasil
}
\keyword{datasets}

