% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outsample.arima.R
\name{outsample.arima}
\alias{outsample.arima}
\title{Previsao fora da amostra com arima}
\usage{
outsample.arima(x, h, k)
}
\arguments{
\item{x}{serie a ser prevista}

\item{h}{horizonte de previsao}

\item{k}{numero de previsoes fora da amostra}
}
\value{
uma lista com
fcast(ts): previsao pontual e intervalos de predicao usando arima 
outdate(num): vetor contendo periodo fora da amostra 
fit(Arima): ultimo modelo arima estimado
}
\description{
outsample.arima e usada para gerar previsoes fora da amostra
usando modelo arima estimado automaticamente segundo package 
\code{forecast}
}

